Mejoras 15-05-2017

+++ src/app/core/services/auth.service.ts
@@ -1,28 +1,45 @@
 import { Injectable } from '@angular/core';
 
+import { environment } from './../../../environments/environment';
 
 @Injectable()
 export class AuthService {
 
   constructor() { console.log('AuthService instance...'); }
 
-  loggIn(user){
-    localStorage.setItem('id_session', user);
-    
-  }
+  logIn(user:string): boolean {
 
-  logOut(){
-    localStorage.removeItem('id_session');
-    localStorage.clear();
-  }
-
-  isLoggedIn(){
-    
-    if ( localStorage.getItem('id_session') != null && localStorage.getItem('id_session') != '' )
+    if ( user !== null && user !== '' && environment.sessionName !== null && environment.sessionName !== '' )
     {
+      localStorage.setItem(environment.sessionName, user);
       return true;
     }
     return false;
+
+  }
+
+  logOut(): boolean {
+
+    if ( environment.sessionName !== null && environment.sessionName !== '' )
+    {
+      localStorage.removeItem(environment.sessionName);
+      localStorage.clear();
+      return true;
+    }
+    return false;
+
+  }
+
+  isLoggedIn(): boolean {
+
+    if ( environment.sessionName !== null && environment.sessionName !== '' )
+    {
+      if ( localStorage.getItem(environment.sessionName) !== null && localStorage.getItem(environment.sessionName) !== '' ){
+        return true;
+      }
+    }
+    return false;
+
   }
 
 }
